Basic outline of memory, reading lines.
This commit outlines a basic structure for the virtual cpu. Instead of
a compiler and cpu, we're going to start with a cpu that also
functions as an interpreter for the assembly language I'm going to
define.

It will translate assembler into opcodes and load them into memory,
line by line.