Virtual CPU can decode operators and operands.
The state of the cpu is such that it can take lines stdin and
decode the instruction (operator) and the two arguments (operand1 &
operand2). It will emit an opcode, and the integer literal or
dereferenced register operands on stdout.

The current goal is a 12-bit architecture, with instruction, operand
flags, and other flags stored in the first of a 3-tuple of 12-bit
operator, and two 12-bit operands stored in the other two slots.

The next step is to store the decoded program into
memory.